extra payout.destination query if not expanded
If we try to sync a payout from a webhook, we will encounter an error because the destination (as a BankAccount or Card) will fail to have the proper customer or stripe_account associated with it. This re-fetches the payout with the destination expanded, so it has the proper context for syncing.